Move Delegations to its own component to avoid unnecessary rerenders
